moved model-utils.js to utils directory
